chore: move ci commands into tools
