+++ .github/workflows/staging-deploy.yml
@@ -504,7 +504,7 @@ jobs:
         env:
           GH_TOKEN: ${{ github.token }}
         run: |
-          python scripts/github_workflow_timing_summary.py \
+          python tools/ci/github_workflow_timing_summary.py \
             --repo "${{ github.repository }}" \
             --run-id "${{ github.run_id }}" \
             --title "Post-merge Timing Summary" \
