+++ docs/project/EPIC-008.testing-strategy.md
@@ -97,6 +97,7 @@ E2E coverage is measured across three tiers of increasing fidelity:
 - **AC8.13.55**: Post-merge staging deploys only for runtime, deploy, E2E, staging workflow, toolchain, or infra-submodule changes.
 - **AC8.13.56**: Coverage and CI metric command entry points run from `tools/` while shared implementations live under `common/`.
 - **AC8.13.57**: SSOT and AC command entry points run from `tools/ssot/` while shared implementations live under `common/ssot/`.
+- **AC8.13.58**: CI and toolchain command entry points run from `tools/ci/` while shared implementations live under `common/ci/`.
 
 Current test and AC coverage status is generated, not hand-maintained here.
 Use `docs/analysis/test-ac-coverage-report.md`,
@@ -425,6 +426,7 @@ These scenarios represent the "Vertical Slices" of user value.
 | AC8.13.55 | Post-merge staging deploys only for runtime, deploy, E2E, staging workflow, toolchain, or infra-submodule changes | `test_AC8_13_55_*` | `scripts/tests/test_ci_change_classifier.py`, `scripts/tests/test_post_merge_e2e_gates.py` | P0 |
 | AC8.13.56 | Coverage and CI metric command entry points run from `tools/` while shared implementations live under `common/` | `test_AC8_13_56_*` | `scripts/tests/test_common_tooling_modules.py`, `scripts/tests/test_ci_metrics_contract.py`, `scripts/tests/test_coverage_policy.py`, `scripts/tests/test_build_unified_lcov.py` | P0 |
 | AC8.13.57 | SSOT and AC command entry points run from `tools/ssot/` while shared implementations live under `common/ssot/` | `test_AC8_13_57_*` | `scripts/tests/test_common_tooling_modules.py`, `scripts/tests/test_ci_metrics_contract.py`, `scripts/tests/test_post_merge_e2e_gates.py` | P0 |
+| AC8.13.58 | CI and toolchain command entry points run from `tools/ci/` while shared implementations live under `common/ci/` | `test_AC8_13_58_*` | `scripts/tests/test_common_tooling_modules.py`, `scripts/tests/test_toolchain_contract.py`, `scripts/tests/test_ci_change_classifier.py`, `scripts/tests/test_github_workflow_timing_summary.py`, `scripts/tests/test_post_merge_e2e_gates.py` | P0 |
 
 **Traceability Ownership**:
 - This table owns the intended AC-to-proof mapping for EPIC-008.
