+++ scripts/dev_backend.py
@@ -15,7 +15,6 @@ import os
 import signal
 import subprocess
 import sys
-import time
 import tomllib
 from pathlib import Path
