+++ scripts/tests/test_github_workflow_timing_summary.py
@@ -1,16 +1,10 @@
-import importlib.util
 import json
 import sys
 from pathlib import Path
 
+sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
-MODULE_PATH = Path(__file__).resolve().parents[1] / "github_workflow_timing_summary.py"
-SPEC = importlib.util.spec_from_file_location("github_workflow_timing_summary", MODULE_PATH)
-assert SPEC is not None
-timing_summary = importlib.util.module_from_spec(SPEC)
-assert SPEC.loader is not None
-sys.modules["github_workflow_timing_summary"] = timing_summary
-SPEC.loader.exec_module(timing_summary)
+from common.ci import github_workflow_timing_summary as timing_summary  # noqa: E402
 
 format_duration = timing_summary.format_duration
 format_timing_summary = timing_summary.format_timing_summary
@@ -66,7 +60,9 @@ def test_AC8_13_34_workflow_timing_summary_reports_queue_and_job_durations() -> 
     assert "| `Classify Changes` | `success` | `4s` |" in summary
 
 
-def test_AC8_13_34_workflow_timing_summary_handles_pending_jobs_without_completed_times() -> None:
+def test_AC8_13_34_workflow_timing_summary_handles_pending_jobs_without_completed_times() -> (
+    None
+):
     run = {
         "url": "https://github.example/actions/runs/2",
         "createdAt": "2026-05-20T10:00:00Z",
@@ -94,7 +90,10 @@ def test_AC8_13_34_workflow_timing_summary_handles_pending_jobs_without_complete
     assert "- Queue delay: `1m 0s`" in summary
     assert "- Execution window: `1m 0s`" in summary
     assert "Longest completed job" not in summary
-    assert "| `Pending Backend` | `in_progress` | `n/a` | `2026-05-20T10:01:00Z` | `n/a` |" in summary
+    assert (
+        "| `Pending Backend` | `in_progress` | `n/a` | `2026-05-20T10:01:00Z` | `n/a` |"
+        in summary
+    )
     assert "| `Queued Backend` | `queued` | `n/a` | `n/a` | `n/a` |" in summary
 
 
@@ -102,7 +101,9 @@ def test_AC8_13_34_load_run_invokes_gh_run_view(monkeypatch) -> None:
     calls = []
 
     class Result:
-        stdout = json.dumps({"url": "https://github.example/actions/runs/3", "jobs": []})
+        stdout = json.dumps(
+            {"url": "https://github.example/actions/runs/3", "jobs": []}
+        )
 
     def fake_run(args, *, check, text, capture_output):
         calls.append((args, check, text, capture_output))
