+++ scripts/tests/test_toolchain_contract.py
@@ -7,9 +7,9 @@ from pathlib import Path
 
 import pytest
 
-sys.path.insert(0, str(Path(__file__).parent.parent))
+sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
-import check_toolchain_contract as contract  # noqa: E402
+from common.ci import check_toolchain_contract as contract  # noqa: E402
 
 
 ROOT = Path(__file__).resolve().parents[2]
